+++ db/migrate/20230121012016_amenity_tag_number_is_not_required.rb
@@ -1,0 +1,5 @@
+class AmenityTagNumberIsNotRequired < ActiveRecord::Migration[7.0]
+  def change
+    change_column_null :amenities, :tag_number, true
+  end
+end